[v2.3.x] prov/efa: Fix cq data size in efa-rdm pkt post
commit 10ca04b1 introduced a bug that introduces an
16 bit cq data intermediate variable when passing to
device. uint16_t is too small and
will silently corrupt when application is using more
than 16 bits while efa provider supports 32 bits.

This patch fixes it.

(cherry picked from commit c40337b45251cf607f9719b2b2b67fdcb74bd2f9)

+++ prov/efa/src/rdm/efa_rdm_pke.c
@@ -411,7 +411,7 @@ ssize_t efa_rdm_pke_sendv(struct efa_rdm_pke **pkt_entry_vec,
 	int ret = 0, pkt_idx, iov_cnt;
 	bool use_inline;
 	uint64_t flags_in_loop;
-	uint16_t cq_data = 0;
+	uint64_t cq_data = 0;
 	uint32_t qpn, qkey;
 
 	assert(pkt_entry_cnt);
@@ -608,7 +608,7 @@ int efa_rdm_pke_write(struct efa_rdm_pke *pkt_entry)
 	int err = 0;
 	struct efa_ah *ah;
 	uint32_t qpn, qkey;
-	uint16_t cq_data = 0;
+	uint64_t cq_data = 0;
 
 	ep = pkt_entry->ep;
 	assert(ep);
